Move checkForLualibType to identifier
